user profile page+tba edit profile

+++ web/src/main/java/com/mannyHelp/web/dto/UsersDto.java
@@ -7,7 +7,7 @@ import lombok.*;
 @NoArgsConstructor
 @AllArgsConstructor
 public class UsersDto {
-    private int userid;
+    private Long userid;
     private String username;
     private String photourl;
     private boolean userOrProvider;
